// projects/codx-ac/src/lib/models/ItemSeries.model.ts
export class ItemSeries{
    recID: any;
    itemID: any;
    seriNo: any;
    seriName: any;
    idiM0: any;
    idiM1: any;
    idiM2: any;
    idiM3: any;
    idiM4: any;
    idiM5: any;
    idiM6: any;
    idiM8: any;
    idiM9: any;
    umid: any;
    conversion: any;
    qty: any;
    cwUM: any;
    cwQty: any;
    manufacturer: any;
    manufaturedDate: any;
    registeredDate: any;
    warrantyDate: any;
    salesWarranty: any;
    extendOn: any;
    statusCodeID: any;
    available: any;
    status: any;
    statusDate: any;
    extendInfo: any;
    customerID: any;
    orderNo: any;
    refID: any;
    sessionID: any;
    barcode: any;
    stop: any;
    note: any;
    autoCreated: any;
    owner: any;
    buid: any;
    createdOn: any;
    createdBy: any;
    modifiedOn: any;
    modifiedBy: any;
}